requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yt-dlp==2023.11.16
anthropic>=0.8.1
python-dotenv==1.0.0
pydantic==2.5.2
python-multipart==0.0.6 
openai>=1.0.0
browser-cookie3==0.19.1
gunicorn==21.2.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cors==1.0.1 